update alerts system
with some smaller changes
