+++ src/utils/functions/index.ts
@@ -3,9 +3,14 @@ import into_readable_datetime from "./into_readable_datetime";
 import truncate_string from "./truncate_string";
 import local_storage from "./local_storage";
 import * as is from "./is";
+import * as _unwrap from "./unwrap";
 
 const assert = {
     ...is
+};
+
+const unwrap = {
+    ..._unwrap
 };
 
 export {
@@ -14,4 +19,5 @@ export {
     into_readable_datetime,
     local_storage,
     assert,
+    unwrap,
 };
